Allow fractional mipmap levels for environment map (uses linear filtering between mipmap levels).  Experiments with dynamic mipmaps based on d omega_i - dosen't look substantially better than static mipmap levels so turned off for now.  Fix some bugs with model transformations when using a "view matrix override" (i.e. when doing a "resample" export request).

+++ src/tetzlaff/models/ReadonlyEnvironmentModel.java
@@ -11,7 +11,7 @@ public interface ReadonlyEnvironmentModel
     boolean isEnvironmentMappingEnabled();
     Vector3 getEnvironmentColor();
     Matrix4 getEnvironmentMapMatrix();
-    int getEnvironmentMapFilteringBias();
+    float getEnvironmentMapFilteringBias();
 
     BackgroundMode getBackgroundMode();
     float getBackgroundIntensity();
