+++ src/tetzlaff/models/ReadonlyLightingModel.java
@@ -15,7 +15,7 @@ public interface ReadonlyLightingModel
     Vector3 getAmbientLightColor();
     boolean isEnvironmentMappingEnabled();
     Matrix4 getEnvironmentMapMatrix();
-    int getEnvironmentMapFilteringBias();
+    float getEnvironmentMapFilteringBias();
 
     ReadonlyLightPrototypeModel getLightPrototype(int i);
     Matrix4 getLightMatrix(int i);
